requirements.txt
langchain
langchain-google-genai
selenium
python-dotenv
pydantic
streamlit